backend/requirements.txt
fastapi
uvicorn
korail2
python-dotenv
pydantic
